Rename components to use Uppercase names
