+++ src/pages/404.tsx
@@ -4,7 +4,7 @@ import CardContent from '@material-ui/core/CardContent';
 import Link from "@material-ui/core/Link"
 import Typography from '@material-ui/core/Typography';
 
-import CenterLayout from "../layouts/center"
+import CenterLayout from "../layouts/Center"
 import SEO from "../components/SEO"
 
 const ConcernedSVG = require("../assets/concerned.svg")
